+++ internal/entity/invite.go
@@ -4,7 +4,7 @@ import "time"
 
 type Invite struct {
 	ID          int32      `json:"id"`
-	UserID      int32      `json:"userId"`
+	UserID      int64      `json:"userId"`
 	InviteHash  string     `json:"inviteHash"`
 	IsActivated bool       `json:"isActivated"`
 	CreatedAt   time.Time  `json:"createdAt"`
