+++ internal/entity/post.go
@@ -4,7 +4,7 @@ import "time"
 
 type Post struct {
 	ID          int32      `json:"id"`
-	UserID      int32      `json:"userId"`
+	UserID      int64      `json:"userId"`
 	User        *User      `json:"user,omitempty"`
 	Title       string     `json:"title"`
 	Short       string     `json:"short"`
